requirements.txt
scipy==1.3.1
numpy==1.17.1
quadprog==0.1.6
seaborn==0.9.0
docutils==0.14
pandas==0.24.2
plotly==3.4.2
tensorflow_gpu==1.12.0
matplotlib==3.0.3
faiss==1.5.3
MulticoreTSNE==0.1
progressbar33==2.4
python_igraph==0.8.0
reportlab==3.5.42
scikit_learn==0.22.2.post1
tensorflow==2.1.0
